Fix(api) : 개인 임계치 수정 요청 DTO Integer -> Long

+++ src/main/java/com/pooli/line/domain/dto/request/UpdateIndividualThresholdReqDto.java
@@ -13,7 +13,7 @@ public class UpdateIndividualThresholdReqDto {
 
     @NotNull @NotEmpty
     @Schema(description = "개인 데이터 임계치(MB)", example = "3000")
-    private Integer individualThreshold;
+    private Long individualThreshold;
 
     @NotNull @NotEmpty
     @Schema(description = "임계치 활성화 여부", example = "true")
